modify

+++ lightning_attn_cute/test.py
@@ -81,9 +81,9 @@ def lightning_attn_no_decay(
     for i in range(NUM_BLOCK):
         si = i * BLOCK
         ei = min(si + BLOCK, N)
-        qi = q[:, :, si:ei, :].contiguous().to(torch.float16)
-        ki = k[:, :, si:ei, :].contiguous().to(torch.float16)
-        vi = v[:, :, si:ei, :].contiguous().to(torch.float16)
+        qi = q[:, :, si:ei, :].contiguous().to(torch.float32)
+        ki = k[:, :, si:ei, :].contiguous().to(torch.float32)
+        vi = v[:, :, si:ei, :].contiguous().to(torch.float32)
 
         qkv_none_diag = torch.matmul(qi, kv.to(qi.dtype)).to(torch.float32)
         o_inter_output[i] = qkv_none_diag.detach().clone()
